+++ src/settings/StaffSlips/components/EditSections/index.js
@@ -1,0 +1,2 @@
+export { default as StaffSlipAboutSection } from './StaffSlipAboutSection';
+export { default as StaffSlipTemplateContentSection } from './StaffSlipTemplateContentSection';