+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change history for ui-circulation
 
+## IN PROGRESS
+* Add metadata information to view of Staff Slips (Settings > Circulation > Staff Slips). Refs UICIRC-855.
+* Implement General information accordion. Refs UICIRC-909
+* Implement Template content accordion. Refs UICIRC-910
+
 ## [8.0.0](https://github.com/folio-org/ui-circulation/tree/v8.0.0) (2023-02-22)
 [Full Changelog](https://github.com/folio-org/ui-circulation/compare/v7.2.1...v8.0.0)
 
@@ -20,7 +25,6 @@
 * An error for spacebar before symbols in "Subject" field when create Patron notice template. Fixes UICIRC-898.
 * UI tests replacement with RTL/Jest for component `RulesEditor`. Refs UICIRC-827.
 * UI tests replacement with RTL/Jest for `Rules-hint`. Refs UICIRC-831.
-* Add metadata information to view of Staff Slips (Settings > Circulation > Staff Slips). Refs UICIRC-855.
 
 ## [7.2.1](https://github.com/folio-org/ui-circulation/tree/v7.2.1) (2022-11-29)
 [Full Changelog](https://github.com/folio-org/ui-circulation/compare/v7.2.0...v7.2.1)